+++ CHANGELOG.md
@@ -11,6 +11,12 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 - ci: `-slim` and `-pytezos` Docker images are no longer published.
 - ci: Docker images no longer contain git, poetry and install_dependencies script.
 
+## [Unreleased]
+
+### Fixed
+
+- config: Fixed crash when `token_transfer` index has `from` or `to` filter.
+
 ## [6.5.2] - 2023-03-09
 
 ### Fixed